% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyHeatwave.R
\name{get.runs.start.end}
\alias{get.runs.start.end}
\title{Get the start of periods with temperature lower than "lessThan" for at least
min.run days long}
\usage{
get.runs.start.end(temperature, lessThan, min.run = 3)
}
\description{
Get the start of periods with temperature lower than "lessThan" for at least
min.run days long
}
