% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyHeatwave.R
\name{get.threshold}
\alias{get.threshold}
\title{Get the three threshold Spic, Sdeb, Sint}
\usage{
get.threshold(temperature, target.quantiles = c(0.995, 0.975, 0.95))
}
\arguments{
\item{temperature}{a vector of temperature}

\item{target.quantiles}{quantile level to compute the three thresholds.
default are 0.995, 0.975, 0.95}
}
\value{
a named list with keys Spic, Sdeb, Sint
}
\description{
Get the three threshold Spic, Sdeb, Sint
}
