% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyHeatwave.R
\name{heatwave.start.end}
\alias{heatwave.start.end}
\title{Get the heatwave statistics}
\usage{
heatwave.start.end(df, thresholds)
}
\arguments{
\item{df}{a data frame, containing a "temperature" column and a "day" column}

\item{thresholds}{three thresholds.}
}
\value{
a data frame with start and end time for heatwave
}
\description{
Get start, end, duration, maximum temperature, global intensity of heatwaves
}
